backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.30.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
numpy==2.3.5
scikit-learn==1.5.2
python-multipart==0.0.6
psycopg2-binary==2.9.10









